+++ src/imio/dataexchange/db/mappers/request.py
@@ -4,21 +4,32 @@ from imio.dataexchange.db import DeclarativeBase
 from imio.dataexchange.db.base import MapperBase
 from sqlalchemy import Boolean
 from sqlalchemy import Column
-from sqlalchemy import Text
 from sqlalchemy import DateTime
+from sqlalchemy import Index
+from sqlalchemy import Text
 from sqlalchemy import func
 
 
 class Request(DeclarativeBase, MapperBase):
-    __tablename__ = u'request'
+    __tablename__ = u"request"
+    __table_args__ = (
+        Index(
+            "request_response_idx",
+            "internal_uid",
+            "response",
+            "expiration_date",
+            "date",
+        ),
+    )
 
-    uid = Column(u'uid', Text, primary_key=True, unique=True, nullable=False)
+    uid = Column(u"uid", Text, primary_key=True, unique=True, nullable=False)
 
-    date = Column(u'date', DateTime, nullable=False, server_default=func.now())
+    internal_uid = Column(u"internal_uid", Text, nullable=False)
 
-    expiration_date = Column(u'expiration_date', DateTime)
+    date = Column(u"date", DateTime, nullable=False, server_default=func.now())
 
-    response = Column(u'response', Text, nullable=True)
+    response = Column(u"response", Text, nullable=True)
 
-    expired = Column(u'expired', Boolean, nullable=False,
-                     server_default='false')
+    expiration_date = Column(u"expiration_date", DateTime)
+
+    expired = Column(u"expired", Boolean, nullable=False, server_default="false")